Добавлено: account_schema, account_service, account_router. добавлена логика для создания аккаунта с привязкой к пользователю по JWT sub заголовку, добавлена проверка в логине на отсутствие пользователя при проверке пароля, исправлены неверные импорты, исправлены критические ошибки: float в models заменен на Numeric, int в schemas заменен на float

+++ app/models/account_model.py
@@ -1,6 +1,8 @@
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 from sqlalchemy import String, Integer, ForeignKey, DECIMAL, Numeric
 
+from decimal import Decimal
+
 from app.core.database import Base
 
 
@@ -10,7 +12,7 @@ class AccountModel(Base):
 
     id: Mapped[int] = mapped_column(primary_key=True)
     user_id: Mapped[int] = mapped_column(ForeignKey("users.id"))
-    balance: Mapped[float] = mapped_column(Numeric(20, 2), default=0.00)
+    balance: Mapped[Numeric] = mapped_column(Numeric(20, 2), default=0.00)
 
     account_owner: Mapped["UserModel"] = relationship(back_populates="accounts")
